Allow searching for basic dependencies
Basic dependencies, i.e. "app-misc/foo" (note there is no constraint at
the front or version at the end of the string), were being missed when
parsing the pquery dump. To fix this, allow the 'bruteForce' parser to
parse (Either ConstrainedDep BasicDep), where BasicDep is a tuple of
(Category, PkgName).

(Either ConstrainedDep BasicDep) is aliased as ParsedDep for
convenience.

These basic dependencies are only displayed in "matching mode", as they
are irrelevant to "non-matching mode", but are helpful when trying to
find reverse dependencies of a package you are hoping to mask.

Signed-off-by: hololeap <hololeap@protonmail.com>